website/requirements.txt
markdown==3.5.1
jinja2==3.1.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Pillow==10.1.0